--- src/async/stompest/async/__init__.py
@@ -1,1 +1,0 @@
-from stompest.async.client import Stomp
