+++ src/twisted/stompest/twisted/__init__.py
@@ -1,0 +1,1 @@
+from stompest.twisted.client import Stomp
