+++ src/twisted/stompest/twisted/tests/async_client_test.py
@@ -5,8 +5,8 @@ from twisted.internet.protocol import Factory
 from twisted.python import log
 from twisted.trial import unittest
 
-from stompest.async import Stomp
-from stompest.async.listener import SubscriptionListener
+from stompest.twisted import Stomp
+from stompest.twisted.listener import SubscriptionListener
 from stompest.config import StompConfig
 from stompest.error import StompCancelledError, StompConnectionError, StompProtocolError
 from stompest.protocol import StompSpec, StompFrame
