+++ src/twisted/stompest/twisted/tests/async_util_test.py
@@ -4,7 +4,7 @@ from twisted.internet import defer, reactor
 from twisted.internet.defer import CancelledError
 from twisted.trial import unittest
 
-from stompest.async.util import InFlightOperations
+from stompest.twisted.util import InFlightOperations
 from stompest.error import StompCancelledError
 
 logging.basicConfig(level=logging.DEBUG)
